# src/test/resources/Collection_Service_API_Spec_v1.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "OpenAPI definition",
        "version": "v0"
    },
    "servers": [
        {
            "url": "https://<host>:<port>/cp/collection-service/",
            "description": "Generated server url"
        }
    ],
    "paths": {
...
        "/api/v1/node-profiles/node-lists": {
            "put": {
                "tags": [
                    "node-list-controller"
                ],
                "summary": "Update node lists to a node profile",
                "operationId": "updateNodeListsToProfile",
                "parameters": [
                    {
                        "name": "nodeProfileId",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/NodeListData"
                                }
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Node lists are updated to the profile"
                    },
                    "207": {
                        "description": "Snmp Group or Auth group mismatch occurred for few of the node lists."
                    },
                    "500": {
                        "description": "Error updating the node lists to the node profile."
                    }
                }
            },
            "post": {
                "tags": [
                    "node-list-controller"
                ],
                "summary": "Add node lists to a node profile",
                "operationId": "addNodeListsToProfile",
                "parameters": [
                    {
                        "name": "nodeProfileId",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/NodeListData"
                                }
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Node lists are added to the profile"
                    },
                    "207": {
                        "description": "Snmp Group or Auth group mismatch occurred for few of the node lists."
                    },
                    "500": {
                        "description": "Error adding the node lists to the node profile."
                    }
                }
            },
            "delete": {
                "tags": [
                    "node-list-controller"
                ],
                "summary": "Delete node lists by its ID",
                "operationId": "deleteNodeLists",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "type": "integer",
                                    "format": "int64"
                                }
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "204": {
                        "description": "Node Lists are deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "integer",
                                        "format": "int64"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error deleting the node lists"
                    }
                }
            }
        },
        "/api/v1/node-profiles/node-lists/filteredList": {
            "put": {
                "tags": [
                    "node-list-controller"
                ],
                "summary": "Get a Paginated Node List by node profile ID",
                "operationId": "getPaginatedSortedNodeList",
                "parameters": [
                    {
                        "name": "nodeProfileId",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "pageNumber",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "sortColumn",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "isAscOrder",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/FilterCriteria"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Found the Node Lists",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NodeListDataPages"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error getting the node lists belonging to the node profile."
                    }
                }
            }
        },
        "/api/v1/node-profiles/node-lists/count": {
            "put": {
                "tags": [
                    "node-list-controller"
                ],
                "summary": "Get the count of Node List by node profile ID",
                "operationId": "getNodeListCount",
                "parameters": [
                    {
                        "name": "nodeProfileId",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/FilterCriteria"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Node Lists count returned",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FilterCriteria"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error getting the node lists count belonging to the node profile."
                    }
                }
            }
        },
...
        "/api/v1/node-profiles/node-lists/import": {
            "post": {
                "tags": [
                    "node-list-controller"
                ],
                "summary": "Import node lists to a node profile using the CSV file",
                "operationId": "importNodeListsToProfile",
                "parameters": [
                    {
                        "name": "nodeProfileId",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "csvFile"
                                ],
                                "type": "object",
                                "properties": {
                                    "csvFile": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                }
                            }
                        },
                        "application/json": {
                            "schema": {
                                "required": [
                                    "csvFile"
                                ],
                                "type": "object",
                                "properties": {
                                    "csvFile": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Node lists are added to the profile"
                    },
                    "207": {
                        "description": "Snmp Group or Auth group mismatch occurred for few of the node lists."
                    },
                    "400": {
                        "description": "Invalid CSV file format."
                    },
                    "500": {
                        "description": "Error adding the node lists to the node profile."
                    }
                }
            }
        },
...
        "/api/v1/networks/collectors/agents/status/{id}": {
            "get": {
                "tags": [
                    "agent-controller"
                ],
                "summary": "Get status of the agent identified by its ID",
                "operationId": "getStatus",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The agent status is successfully retrieved.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AgentData"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Agent not found."
                    },
                    "500": {
                        "description": "Error getting the agent status"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "NodeListData": {
                "required": [
                    "nodeIp"
                ],
                "type": "object",
                "properties": {
                    "nodeIp": {
                        "type": "string"
                    },
                    "nodeManagementIp": {
                        "type": "string"
                    },
                    "authGroupName": {
                        "type": "string"
                    },
                    "snmpGroupName": {
                        "type": "string"
                    },
                    "snmpGroupMismatch": {
                        "type": "boolean",
                        "readOnly": true
                    },
                    "authGroupMismatch": {
                        "type": "boolean",
                        "readOnly": true
                    },
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
...
            "FilterCriteria": {
                "type": "object",
                "properties": {
                    "columnName": {
                        "type": "string"
                    },
                    "columnValue": {
                        "type": "string"
                    }
                }
            },
            "NodeListDataPages": {
                "type": "object",
                "properties": {
                    "count": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "nodeLists": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/NodeListData"
                        }
                    }
                }
            },
...
            "AgentData": {
                "required": [
                    "name",
                    "params",
                    "type"
                ],
                "type": "object",
                "properties": {
                    "updateDate": {
                        "type": "string",
                        "format": "date",
                        "readOnly": true
                    },
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "SR_PCE_AGENT",
                            "NETFLOW_AGENT",
                            "TRAFFIC_POLLER"
                        ]
                    },
                    "params": {
                        "type": "string"
                    }
                }
            },
...
        }
    }
}